frontend: add documentation
Going forward, this code will be refactored. However, in order to fix and
upgrade the code we must first be able to understand it :')

This commit refactors the code so that variable names and general organisation
of the code is easy to follow.

We also add extensive comments and documentation, so it is obvious what bit of
code does what with a simple glance.