corna->subdomain.py: Add full post URI
Previously, our single post object did not contain the
full post URI to be used on the client. This meant
that the client would have to build the href from
separate components of the post object e.g. domain_name.

This commit simplifies the work done client side by passing
the full post URI down to along with the post object.