[bug]posts: Do not send optional fields
We've seen issues where the client sends a `null` value to the posts endpoint
with a field which is optional.

Although, it is fine to completely leave out the optional fields sending them
with `null` or with empty strings can cause marshmallow validation errors.

This commit fixes the bug with two methods: leave out optional fields on the
client; remove `null` values from the backend.

We need to ensure the backend work is done as there is a zero-trust contract
with the client i.e. we cannot guarantee that optional fields will indeed be
left out of a given API call.