frontend/characters.ts: process dynamic htmx DIVs
Currently, we cannot interact with the character divs in on the corna card as
they are dynamically created using javascript. This is a known problem[1][2],
htmx does not pick up content that has been created outside of the regular
request/response cycle.

This commit ensures that htmx process the dynamically created html. To do this
we also need to install htmx via npm so it can be used in our scripts -
in our main html pages we use the CDN link in a script tag, which makes it
inaccessible at build time.

[1] https://htmx.org/api/#process
[2] https://stackoverflow.com/a/73659697