CI: bump eslint runtime environment
For some reason, we didn't bump `eslint` during our python 3.12 upgrade. The
only reason it was `ubuntu-20.04` to begin with was to keep consistency with
the rest of the CI runners. Now GHA's is deprecating `ubuntu-20.04`[1], its as
good time as any to upgrade.

This commit bumps the runtime version to `ubuntu-latest` which matches the rest
of our CI runners.

[1] https://github.com/actions/runner-images/issues/11101