corna->media.py: chunked upload
Currently, the client sends the full file to our server in one request. This
is problematic because the bottleneck is our pre-configured nginx `max_buffer_size`
(which sits at 20MB).

In order to give the backend service control over the max file sizing, we have
opted to expose a chunked upload option to clients.

This commit adds the functionality to allow clients to send a range of small
chunks to the server rather than attempted to send the entire file. The core
tracking method used is the user generated uploadId and the matadata file which
is created for each upload and tracks how many chunks the backend has received
vs how many are expected. This allows clients to continue uploads in the case of
failures and/or network drops.

NOTE: this is a first implementation. This code is NOT thread safe, and there are
significant flaws if two processes try to write to the metadata file at the same
time. This will be fixed in due time.