corna->media.py: merge chunks
This commit adds an implementation to merge chunks on the server once they
have all be uploaded.

Once merged, the complete file is saved using the normal backend media saving
process and gets a custom url generated.

The reason this is a separate endpoint is to allow clients full control of when
the file gets merged and also reduces the complexity of code as we now do not
need various branching cases for different states.

The functionality here implements as very shoddy locking mechanism so no two
processes can start merging the same file. Future improvements will make this more
robust. But hey, its better than the upload endpoint :')