meta.py: return path from `mkdir`
Turns out we were not returning a path from `mkdir`. The original implementation
was under the assumption the the original variable (in the caller) would become
a path object but python is pass by value not pass by reference (loosely speaking).

To rectify this, this commit updates the `mkdir` function to return the newly
created path.