frontend->media.ts: add client support for chunking
This commit updates the client logic so it takes advantage of the new chunking
capabilities.

We still want to use the legacy mechanisms for files that are smaller than our
nginx `max_buffer_size`.

This is generally a poor implementation as it does not handle errors well or
implement retry mechanisms. But it's a step in the right direction.