tests->utils.py: invalid escape error
One of the test cases for ensuring our in coming HTML is properly
escaped and cleaned in causing a warning locally and an error in
upstream CI.

The crux of this is that its, understandably, weird python string syntax.
Due to the fact we're testing for incoming, client strings it makes sense
python is unhappy.

This commit adds a quick fix but turning the string in the test cases into
raw strings rather than python strings.