corna->media.py: add support for calculating aspect ratios
During rendering we've seen HTML creating arbitary DIV sizes (based on
hardcoded values in the markup) which make images and videos look really weird.

To resolve this we would like to pass asset sizes down to the client so media
assets can be rendered in the correct aspect ratio. This also has the added
benefit of allowing us to filter placeholder images, that come standard with the
service, by aspect ratio so clients can have assets that match the theme
aesthetic.

This commit adds support for saving and managing asset dimensions and aspect
ratios. It also adds a bunch of plumbing to make this all work.